backend/requirements.txt
asgiref==3.9.2
black==25.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
cryptography==46.0.1
defusedxml==0.7.1
Django==4.2.24
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
flake8==7.3.0
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.7.0
isort==6.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
pytokens==0.1.10
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
